% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetProt.R
\docType{data}
\name{CR}
\alias{CR}
\title{This is the CR colorectal cancer data to be included in my package}
\references{
\url{http://www.nature.com/nm/journal/v21/n4/full/nm.3807.html}
}
\author{
Wilson Goh \email{goh.informatics@gmail.com}
}
\keyword{data}
