% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetProt.R
\name{essnet}
\alias{essnet}
\title{Based on the output of internal_substraction, this calculates a p-value for each complex}
\usage{
essnet(diff_matrix, complex_matrix, overlap_threshold)
}
\arguments{
\item{diff_matrix}{A matrix of deltas derived from internal_substraction}

\item{overlap_threshold}{The required minimal overlap between the proteomics data and the network}

\item{complete_matrix}{The original expression matrix}
}
\value{
p_val_vec A vector of p-values, one for each complex
}
