% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetProt.R
\name{fsnet}
\alias{fsnet}
\title{Based on the rank weight matrix and the list of complexes, give a matrix of scores per complex}
\usage{
fsnet(rank_weight_matrix, complex_matrix, overlap_threshold)
}
\arguments{
\item{rank_weight_matrix}{A matrix of weighted ranks}

\item{complex_matrix}{The original expression matrix}

\item{overlap_threshold}{The required minimal overlap between the proteomics data and the network}
}
\value{
output A list of two matrices of fsnet scores, x vs y, and y vs x, for complexes for all samples
}
