% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetProt.R
\name{generate_proteomics_sim}
\alias{generate_proteomics_sim}
\title{Proteomics data simulation function}
\usage{
generate_proteomics_sim(data, fact, prop, rounds, effect_size)
}
\arguments{
\item{data}{A one class data matrix}

\item{fact}{Factors}

\item{prop}{Proportion of variables to insert effect size}

\item{rounds}{Number ofsimulations}

\item{effect_size}{A vector of intended effect sizes}
}
\value{
prop A set of simulated datasets
}
