% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetProt.R
\name{generate_pseudo_cplx}
\alias{generate_pseudo_cplx}
\title{Pseudo-complex simulation function}
\usage{
generate_pseudo_cplx(tdata, fact, purity)
}
\arguments{
\item{tdata}{A one class data matrix}

\item{fact}{Factors}

\item{purity}{Proportion of differential proteins in complex}
}
\value{
prop A set of simulated pseudo-complexes
}
