% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetProt.R
\name{generate_weight_matrix}
\alias{generate_weight_matrix}
\title{Generates a weight matrix by applying the generate weight vector function to all columns}
\usage{
generate_weight_matrix(x)
}
\arguments{
\item{x}{data matrix}
}
\value{
weight_matrix A weight matrix
}
