% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetProt.R
\name{gsea}
\alias{gsea}
\title{The original GSEA algorithm based on the Kolmogorov-Smirnov}
\usage{
gsea(data, complex_vector, factors)
}
\arguments{
\item{data}{expression matrix}

\item{complex_vector}{A list of complexes}

\item{factors}{levels indicating the class of samples in the expression matrix}
}
\value{
gsea_pval A vector of p-values; one for each complex
}
