% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetProt.R
\name{he}
\alias{he}
\title{The hypergeometric enrichment pipeline}
\usage{
he(x, y, z)
}
\arguments{
\item{x}{protein expression matrix}

\item{y}{complex vector}

\item{z}{class factors}
}
\value{
hyp_pval A vector of p-values, one for each complex
}
