% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetProt.R
\name{internal_substraction}
\alias{internal_substraction}
\title{This generates all paired differences, deltas, for each gene between samples in class A and B}
\usage{
internal_substraction(group_A, group_B)
}
\arguments{
\item{group_A}{samples in group A}

\item{group_A}{samples in group B}
}
\value{
all_diffs A matrix of deltas between class A and B
}
