% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetProt.R
\name{ppfsnet_theoretical_t_test}
\alias{ppfsnet_theoretical_t_test}
\title{Uses the original FSNET matrix but applies PPFSNET's calculation method}
\usage{
ppfsnet_theoretical_t_test(x, y)
}
\arguments{
\item{x}{FSNET matrix}

\item{y}{test significance level}
}
\value{
out p-values based on pfsnet
}
