% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetProt.R
\name{qpsp}
\alias{qpsp}
\title{QPSP rank matrix generation function}
\usage{
qpsp(rank_weight_matrix, complex_list)
}
\arguments{
\item{rank_weight_matrix}{A matrix of weights based on ranks}

\item{complex_list}{A list of complexes}
}
\value{
output A vector of p-values
}
