% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetProt.R
\name{qpsp_generate_rank_matrix}
\alias{qpsp_generate_rank_matrix}
\title{QPSP rank matrix generation function}
\usage{
qpsp_generate_rank_matrix(x)
}
\arguments{
\item{x}{expression matrix}
}
\value{
rank_matrix A matrix of weighted ranks
}
