% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetProt.R
\name{snet_generate_weight_matrix}
\alias{snet_generate_weight_matrix}
\title{Generates a matrix of snet weights for a given data matrix}
\usage{
snet_generate_weight_matrix(x)
}
\arguments{
\item{x}{An expression matrix}
}
\value{
output A matrix of weights
}
