% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetProt.R
\name{snet_weight_vector}
\alias{snet_weight_vector}
\title{Generates a vector of snet weights for a given sample}
\usage{
snet_weight_vector(x)
}
\arguments{
\item{x}{An expression vector}
}
\value{
output A vector of weights
}
